+++ Examples/gifeditor/Package.swift
@@ -39,8 +39,6 @@ let package = Package(
   ],
   dependencies: [
     .package(name: "swift-tui", path: "../.."),
-    .package(name: "SwiftTUIArguments", path: "../../Platforms/Arguments"),
-    .package(name: "SwiftTUIWebHost", path: "../../Platforms/WebHost"),
     .package(path: "../../Vendor/swift-gif"),
   ],
   targets: [
@@ -70,8 +68,8 @@ let package = Package(
       dependencies: [
         "GIFEditor",
         .product(name: "SwiftTUI", package: "swift-tui"),
-        .product(name: "SwiftTUIArguments", package: "SwiftTUIArguments"),
-        .product(name: "SwiftTUIWebHostCLI", package: "SwiftTUIWebHost"),
+        .product(name: "SwiftTUIArguments", package: "swift-tui"),
+        .product(name: "SwiftTUIWebHostCLI", package: "swift-tui"),
       ]
     ),
     .testTarget(