+++ Examples/gitviz/Package.swift
@@ -9,8 +9,6 @@ let package = Package(
   ],
   dependencies: [
     .package(name: "swift-tui", path: "../.."),
-    .package(name: "SwiftTUICLI", path: "../../Platforms/CLI"),
-    .package(name: "SwiftTUIArguments", path: "../../Platforms/Arguments"),
   ],
   targets: [
     .executableTarget(
@@ -18,8 +16,8 @@ let package = Package(
       dependencies: [
         .product(name: "SwiftTUI", package: "swift-tui"),
         .product(name: "SwiftTUICharts", package: "swift-tui"),
-        .product(name: "SwiftTUICLI", package: "SwiftTUICLI"),
-        .product(name: "SwiftTUIArguments", package: "SwiftTUIArguments"),
+        .product(name: "SwiftTUICLI", package: "swift-tui"),
+        .product(name: "SwiftTUIArguments", package: "swift-tui"),
       ]
     ),
     .testTarget(